+++ src/pipes/wallet-to-usd/wallet-to-usd.ts
@@ -15,7 +15,7 @@ export class WalletToUSDPipe implements PipeTransform {
      * Takes a value and makes it lowercase.
      */
     transform(wallet: NWallet.WalletItem) {
-        const value = Number.parseInt(wallet.amount) * wallet.price;
+        const value = Number.parseFloat(wallet.amount) * wallet.price;
         const floor = Math.floor(value * 100) / 100;
         return `$${floor} USD`;
     }
